ruby: Change tokens to be a part of an api object instead of global
As more api clients get added in the future, we're going to want to use
different auth tokens. Currently, all api clients use a single global
token, so if you wanted to use both worker and mq, you wouldn't be able
to since both services have seperate tokens. The fix to this problem is
to make token contained to a specific api client. As a result, api call
functions can no longer be static.